+++ sdk/communication/azure-communication-sms/azure/communication/sms/__init__.py
@@ -1,8 +1,16 @@
 from ._sms_client import SmsClient
+from ._delivery_reports_client import DeliveryReportsClient
+from ._opt_outs_client import OptOutsClient
+from ._telco_messaging_client import TelcoMessagingClient
 
-from ._models import SmsSendResult
+from ._models import SmsSendResult, OptOutResult, OptOutCheckResult
 
 __all__ = [
     "SmsClient",
+    "DeliveryReportsClient",
+    "OptOutsClient",
+    "TelcoMessagingClient",
     "SmsSendResult",
+    "OptOutResult",
+    "OptOutCheckResult",
 ]