+++ sdk/communication/azure-communication-sms/azure/communication/sms/_sms_client.py
@@ -11,6 +11,7 @@ from azure.communication.sms._generated.models import (
     SendMessageRequest,
     SmsRecipient,
     SmsSendOptions,
+    MessagingConnectOptions,
 )
 from azure.communication.sms._models import SmsSendResult
 from azure.core.credentials import TokenCredential, AzureKeyCredential
@@ -24,18 +25,33 @@ from ._version import SDK_MONIKER
 class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
     """A client to interact with the AzureCommunicationService Sms gateway.
 
-    This client provides operations to send an SMS via a phone number.
+    This client provides operations to send an SMS via a phone number with support for
+    advanced delivery options including delivery report timeouts and Messaging Connect
+    partner networks.
 
     :param str endpoint:
         The endpoint url for Azure Communication Service resource.
     :param Union[TokenCredential, AzureKeyCredential] credential:
         The credential we use to authenticate against the service.
+    :keyword str api_version:
+        The API version to use for requests. If not specified, the default API version will be used.
+
+    .. admonition:: Example:
+
+        .. literalinclude:: ../samples/send_sms_with_advanced_options_sample.py
+            :start-after: [START send_sms_with_advanced_options]
+            :end-before: [END send_sms_with_advanced_options]
+            :language: python
+            :dedent: 4
+            :caption: Sending SMS with advanced delivery and partner options
     """
 
     def __init__(
             self,
             endpoint,  # type: str
             credential,  # type: Union[TokenCredential, AzureKeyCredential]
+            *,
+            api_version: Optional[str] = None,
             **kwargs  # type: Any
     ):
         # type: (...) -> None
@@ -50,20 +66,31 @@ class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
 
         self._endpoint = endpoint
         self._authentication_policy = get_authentication_policy(endpoint, credential)
+        
+        # If api_version is provided, pass it to the service client
+        service_kwargs = kwargs.copy()
+        if api_version is not None:
+            # The service client configuration will pick up api_version from kwargs
+            service_kwargs['api_version'] = api_version
+            
         self._sms_service_client = AzureCommunicationSMSService(
-            self._endpoint, authentication_policy=self._authentication_policy, sdk_moniker=SDK_MONIKER, **kwargs
+            self._endpoint, authentication_policy=self._authentication_policy, sdk_moniker=SDK_MONIKER, **service_kwargs
         )
 
     @classmethod
     def from_connection_string(
             cls,
             conn_str,  # type: str
+            *,
+            api_version: Optional[str] = None,
             **kwargs  # type: Any
     ):  # type: (...) -> SmsClient
         """Create SmsClient from a Connection String.
 
         :param str conn_str:
             A connection string to an Azure Communication Service resource.
+        :keyword str api_version:
+            The API version to use for requests. If not specified, the default API version will be used.
         :returns: Instance of SmsClient.
         :rtype: ~azure.communication.sms.SmsClient
 
@@ -78,7 +105,7 @@ class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
         """
         endpoint, access_key = parse_connection_str(conn_str)
 
-        return cls(endpoint, AzureKeyCredential(access_key), **kwargs)
+        return cls(endpoint, AzureKeyCredential(access_key), api_version=api_version, **kwargs)
 
     @distributed_trace
     def send(
@@ -89,6 +116,9 @@ class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
             *,
             enable_delivery_report: bool = False,
             tag: Optional[str] = None,
+            delivery_report_timeout_in_seconds: Optional[int] = None,
+            messaging_connect_api_key: Optional[str] = None,
+            messaging_connect_partner_name: Optional[str] = None,
             **kwargs: Any,
     ) -> List[SmsSendResult]:
         """Sends SMSs to phone numbers.
@@ -101,13 +131,40 @@ class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
          message on the Azure Resource EventGrid.
         :keyword str tag: Use this field to provide metadata that will then be sent back in the corresponding
          Delivery Report.
+        :keyword int delivery_report_timeout_in_seconds: Time to wait for a delivery report. After this time a
+         delivery report with timeout error code is generated. Must be between 60 and 43200 seconds.
+        :keyword str messaging_connect_api_key: API key for Messaging Connect Partner. 
+         Must be provided together with messaging_connect_partner_name.
+        :keyword str messaging_connect_partner_name: Partner name for Messaging Connect.
+         Must be provided together with messaging_connect_api_key.
         :return: A list of SmsSendResult.
         :rtype: [~azure.communication.sms.models.SmsSendResult]
+        :raises ValueError: If messaging_connect_api_key is provided without messaging_connect_partner_name or vice versa.
         """
         if isinstance(to, str):
             to = [to]
 
-        sms_send_options = SmsSendOptions(enable_delivery_report=enable_delivery_report, tag=tag)
+        # Validate MessagingConnect fields
+        if (messaging_connect_api_key is None) != (messaging_connect_partner_name is None):
+            raise ValueError(
+                "Both messaging_connect_api_key and messaging_connect_partner_name must be provided together, "
+                "or both must be None."
+            )
+
+        # Create MessagingConnectOptions if both fields are provided
+        messaging_connect = None
+        if messaging_connect_api_key is not None and messaging_connect_partner_name is not None:
+            messaging_connect = MessagingConnectOptions(
+                api_key=messaging_connect_api_key,
+                partner=messaging_connect_partner_name
+            )
+
+        sms_send_options = SmsSendOptions(
+            enable_delivery_report=enable_delivery_report, 
+            tag=tag,
+            delivery_report_timeout_in_seconds=delivery_report_timeout_in_seconds,
+            messaging_connect=messaging_connect
+        )
 
         request = SendMessageRequest(
             from_property=from_,
@@ -123,16 +180,16 @@ class SmsClient(object):  # pylint: disable=client-accepts-api-version-keyword
         )
 
         response = self._sms_service_client.sms.send(
-            request,
+            request.serialize(),
             **kwargs
         )
 
         return [
             SmsSendResult(
-                to=item.to,
-                message_id=item.message_id,
-                http_status_code=item.http_status_code,
-                successful=item.successful,
-                error_message=item.error_message
-            ) for item in response.value
+                to=item["to"],
+                message_id=item["messageId"],
+                http_status_code=item["httpStatusCode"],
+                successful=item["successful"],
+                error_message=item.get("errorMessage")
+            ) for item in response["value"]
         ]